Moved rumors generation to NewTurnProcessor
